refactor: fix person module naming
