+++ src/app.module.ts
@@ -8,7 +8,7 @@ import { DatabaseModule } from './database/database.module';
 import { EmployeesModule } from './employees/employees.module';
 import { databaseModuleOptions } from './config/database-module-options';
 import { AircraftModule } from './aircraft/aircraft.module';
-import { PersonModule } from './person/person.module';
+import { PersonsModule } from './persons/persons.module';
 import { DocumentModule } from './document/document.module';
 import { AirportsModule } from './airports/airports.module';
 import { TerminalsModule } from './terminals/terminals.module';
@@ -22,7 +22,7 @@ import { FlightsModule } from './flights/flights.module';
     DatabaseModule.forRootAsync(databaseModuleOptions),
     EmployeesModule,
     AircraftModule,
-    PersonModule,
+    PersonsModule,
     DocumentModule,
     AirportsModule,
     TerminalsModule,
