+++ src/document/document.module.ts
@@ -2,10 +2,10 @@ import { Module } from '@nestjs/common';
 import { DocumentService } from './document.service';
 import { DocumentController } from './document.controller';
 import { DocumentRepository } from './document.repository';
-import { PersonModule } from 'src/person/person.module';
+import { PersonsModule } from 'src/persons/persons.module';
 
 @Module({
-  imports: [PersonModule],
+  imports: [PersonsModule],
   controllers: [DocumentController],
   providers: [DocumentService, DocumentRepository],
 })
