--- src/person/entities/person.entity.ts
@@ -1,1 +1,0 @@
-export class Person {}
