--- src/person/person.module.ts
@@ -1,11 +1,0 @@
-import { Module } from '@nestjs/common';
-import { PersonService } from './person.service';
-import { PersonController } from './person.controller';
-import { PersonRepository } from './person.repository';
-
-@Module({
-  controllers: [PersonController],
-  providers: [PersonService, PersonRepository],
-  exports: [PersonService],
-})
-export class PersonModule {}
